Application

Documentation

getApplicationDev :: IO (Settings, Application)

For yesod devel, return the Warp settings and WAI Application.

appMain :: IO ()

The main function for an executable running this site.

develMain :: IO ()

main function for use by yesod devel

makeFoundation :: AppSettings -> IO App

This function allocates resources (such as a database connection pool), performs initialization and returns a foundation datatype value. This is also the place to put your migrate statements to have automatic database migrations handled by Yesod.

makeLogWare :: App -> IO Middleware

for DevelMain

getApplicationRepl :: IO (Int, App, Application)

shutdownApp :: App -> IO ()

for GHCI

handler :: Handler a -> IO a

Run a handler

db :: ReaderT SqlBackend (HandlerT App IO) a -> IO a

Run DB queries