AuthSite

Description

Subsite for email-and-passphrase authentication.

Quickstart:

Add AuthSite as a subsite to your routes somewhere.

Mappend migrateAuth to your site's Migration when running migrations.

Have your site datatype instantiate AuthMaster (this is the hard part).

Subsite interface

Integration with your site

data AuthSite

Foundation type for the subsite.

Constructors

AuthSite

migrateAuth :: Migration

class AuthMaster master where

Any site that uses this subsite needs to instantiate this class.

Methods

postLoginRoute :: master -> Route master

Where to go after login

postLogoutRoute :: master -> Route master

Where to go after logout

loginHandler :: HandlerT master IO TypedContent

What to show on the login page. This page should have a form that posts Credentials to 'Route LoginR'. See AuthHarness in the tests for a simplistic example.

createAccountHandler :: HandlerT master IO TypedContent

What to show on the create-account page. This page should post Credentials to CreateAccountR.

resetPassphraseHandler :: HandlerT master IO TypedContent

What to show on the reset-passphrase page. This page should post Credentials to ResetPassphraseR

verifyAccountHandler :: HandlerT master IO TypedContent

What to show on the verify-account page. This page should post Text (the token) to VerifyAccountR

sendAuthEmail :: AuthMailMessage -> AuthEmail -> HandlerT master IO ()

This module sends emails, in case that wasn't obvious. Network.Mail.Mime or Network.Mail.Mime.SES have good options for this method.

type AuthUser = Entity User

This is a cheapo synonym. Eventually AuthUser should be part of the AuthMaster interface, for this module to be usable outside of Snowdrift.

Routes

Haddock seems to dump the entire Route definition, even though the only interesting thing here is the AuthSite instance with LoginR, LogoutR, CreateAccountR, VerifyAccountR, and ResetPassphraseR.

data family Route a

The type-safe URLs associated with a site argument.

Querying auth status

Duplicating the Yesod.Auth interface

maybeAuth :: (YesodPersist m, YesodPersistBackend m ~ SqlBackend) => HandlerT m IO (Maybe AuthUser)

If the user is authenticated, get the corresponding Entity.

requireAuth :: (Yesod m, YesodPersist m, YesodPersistBackend m ~ SqlBackend) => HandlerT m IO AuthUser

If the user is not authenticated, this will cause a redirect to your site's authRoute or simply return notAuthenticated.

Note that Yesod.Auth is smarter about not redirecting during an API request, but we don't support that yet.

Helper for writing handlers

credentialsForm :: (RenderMessage (HandlerSite m) FormMessage, MonadHandler m) => AForm m Credentials

A decent default form for Credentials.

data Credentials

The "email and passphrase" representing the main purpose of this module.

Constructors

Credentials

Fields

credsIdent :: AuthEmail

credsPass :: ClearPassphrase

newtype AuthEmail

Sanity-preserving type

Constructors

AuthEmail

Fields

fromAuth :: Text

newtype ClearPassphrase

Sanity-preserving type

Constructors

ClearPassphrase

Fields

fromClear :: Text

Types

data AuthMailMessage

The type of message you are expected to send.

Constructors

VerifyUserCreation AuthToken
A user is signing up

VerifyPassReset AuthToken
A user wants to reset their passphrase

BadUserCreation
A user tried to sign up with an existing address

BadPassReset
A user tried to reset with an address that doesn't exist

newtype AuthToken

A token used to confirm an email address.

Constructors

AuthToken

Fields

fromAuthToken :: Text

Internals

provisional :: Credentials -> Verification -> IO ProvisionalUser

Generate a provisional user

logout :: Yesod master => HandlerT master IO ()

Log out by deleting the session var

priviligedProvisionalUser :: MonadIO m => Credentials -> SqlPersistT m Verification

Store a provisional user for later verification. Returns the token to use for verification.

privilegedCreateUser :: MonadIO m => VerifiedUser -> SqlPersistT m ()

Insert a new user into the database.

priviligedLogin :: Yesod master => AuthUser -> HandlerT master IO ()

This privileged function must be used with care. It modifies the user's session; it's the difference between being logged in and not!

data VerifiedUser

Internal

Constructors

VerifiedUser

Fields

verifiedEmail :: Text

verifiedDigest :: ByteString

data Verification

Internal

Constructors

Verification

Fields

verifyEmail :: AuthEmail

verifyToken :: AuthToken

persistent-generated types

data ProvisionalUser

Constructors

ProvisionalUser

Fields

provisionalUserEmail :: !Text

provisionalUserDigest :: !ByteString

provisionalUserToken :: !Text

provisionalUserCreationTime :: !UTCTime