EmailAuth

Description

This module defines methods needed for the YesodAuthEmail instance for App.

Documentation

addUnverified :: MonadIO m => Email -> VerKey -> SqlPersistT m VerifyEmailId

Add a new, unverified user. Uniqueness guarantees this blows up if the user already exists.

getVerifyKey :: MonadIO m => VerifyEmailId -> SqlPersistT m (Maybe VerKey)

setVerifyKey :: MonadIO m => VerifyEmailId -> VerKey -> SqlPersistT m ()

verifyAccount :: MonadIO m => VerifyEmailId -> SqlPersistT m (Maybe UserId)

getPassword :: MonadIO m => UserId -> SqlPersistT m (Maybe Text)

setPassword :: MonadIO m => UserId -> SaltedPass -> SqlPersistT m ()

getEmailCreds :: (MonadIO m, AuthEmailId site ~ Key VerifyEmail, AuthId site ~ Key User) => Text -> SqlPersistT m (Maybe (EmailCreds site))

Since we only use email for identification, this function is degenerate in a few ways.

getEmail :: MonadIO m => VerifyEmailId -> SqlPersistT m (Maybe Email)