Handler

Description

Common handler functions.

Documentation

getFaviconR :: Handler TypedContent

getRobotsR :: Handler TypedContent

getHomeR :: Handler Html

Homepage is an introduction to the site for non-logged-in viewers, and the dashboard for logged-in viewers.

getWelcomeR :: Handler Html

getDashboardR :: Handler Html

getHowItWorksR :: Handler Html

getPrivacyR :: Handler Html

getTermsR :: Handler Html

getSearchR :: Handler Html

getProjectsR :: Handler Html

getTrademarksR :: Handler Html

getDonateR :: Handler Html

getAboutR :: Handler Html

getSponsorsR :: Handler Html

getJsLicensesR :: Handler Html

getMerchandiseR :: Handler Html

getSnowdriftProjectR :: Handler Html

For MVP, there is one, hard-coded project: Snowdrift

getSnowdriftWikiSearchR :: Text -> Handler Html

Prevents breakage of external links to the old wiki. See https://tree.taiga.io/project/snowdrift/us/359