Settings

Description

Settings are centralized, as much as possible, into this file. This includes database connection settings, static file locations, etc. In addition, you can configure a number of different aspects of Yesod by overriding methods in the Yesod typeclass. That instance is declared in the Foundation.hs file.

Documentation

data AppSettings

Runtime settings to configure this application. These settings can be loaded from various sources: defaults, environment variables, config files, theoretically even a database.

Constructors

AppSettings

Fields

appStaticDir :: String

Directory from which to serve static files.

appDatabaseConf :: PostgresConf

Configuration settings for accessing the database.

appRoot :: Maybe Text

Base for all generated URLs. If Nothing, determined from the request headers.

appHost :: HostPreference

Host/interface the server should bind to.

appPort :: Int

Port to listen on

appIpFromHeader :: Bool

Get the IP address from the header when logging. Useful when sitting behind a reverse proxy.

appDetailedRequestLogging :: Bool

Use detailed request logging system

appShouldLogAll :: Bool

Should all log messages be displayed?

appReloadTemplates :: Bool

Use the reload version of templates

appMutableStatic :: Bool

Assume that files in the static dir may change after compilation

appSkipCombining :: Bool

Perform no stylesheet/script combining

widgetFileSettings :: WidgetFileSettings

Settings for widgetFile, such as which template languages to support and default Hamlet settings.

For more information on modifying behavior, see:

https://github.com/yesodweb/yesod/wiki/Overriding-widgetFile

combineSettings :: CombineSettings

How static files should be combined.

widgetFile :: String -> Q Exp

configSettingsYmlBS :: ByteString

Raw bytes at compile time of config/settings.yml

configSettingsYmlValue :: Value

config/settings.yml, parsed to a Value.

compileTimeAppSettings :: AppSettings

A version of AppSettings parsed at compile time from config/settings.yml.

combineStylesheets :: Name -> [Route Static] -> Q Exp

combineScripts :: Name -> [Route Static] -> Q Exp